+++ protocols/floodsub/src/lib.rs
@@ -28,6 +28,7 @@ pub mod protocol;
 mod layer;
 mod topic;
 
+#[allow(clippy::derive_partial_eq_without_eq)]
 mod rpc_proto {
     include!(concat!(env!("OUT_DIR"), "/floodsub.pb.rs"));
 }