+++ misc/metrics/CHANGELOG.md
@@ -14,6 +14,8 @@
 
 - Track number of connected nodes supporting a specific protocol via the identify protocol. See [PR 2734].
 
+- Update to `libp2p-core` `v0.35.0`.
+
 [PR 2734]: https://github.com/libp2p/rust-libp2p/pull/2734/
 
 # 0.7.0