+++ misc/metrics/CHANGELOG.md
@@ -12,6 +12,16 @@
 
 - Update to `libp2p-kad` `v0.39.0`.
 
+- Track number of connected nodes supporting a specific protocol via the identify protocol. See [PR 2734].
+
+- Update to `libp2p-core` `v0.35.0`.
+
+- Update to `prometheus-client` `v0.17.0`. See [PR 2761].
+
+[PR 2761]: https://github.com/libp2p/rust-libp2p/pull/2761/
+
+[PR 2734]: https://github.com/libp2p/rust-libp2p/pull/2734/
+
 # 0.7.0
 
 - Update to `libp2p-core` `v0.34.0`.