+++ misc/metrics/CHANGELOG.md
@@ -16,9 +16,9 @@
 
 - Update to `libp2p-core` `v0.35.0`.
 
-- Update to `prometheus-client` `v0.17.0`. See [PR 2761].
+- Update to `prometheus-client` `v0.18.0`. See [PR 2822].
 
-[PR 2761]: https://github.com/libp2p/rust-libp2p/pull/2761/
+[PR 2822]: https://github.com/libp2p/rust-libp2p/pull/2761/
 
 [PR 2734]: https://github.com/libp2p/rust-libp2p/pull/2734/